requirements.txt
psycopg
psycopg2
Pillow
tailwind
django-tailwind
django-browser-reload
django-alpine
django-redis
pymongo
django-seed
django-altcha
django-cors-headers
tqdm
htmx